backend/requirements.txt
flask
flask-cors
numpy
opencv-python-headless
torch
Pillow
git+https://github.com/facebookresearch/sam2.git